feat: Update app.js to add routes for retrieving all blogs and a single blog
This commit modifies the app.js file to add two new routes: "/all-blogs" and "/single-blog". The "/all-blogs" route retrieves all blog posts from the database and sends them as a response. The "/single-blog" route retrieves a specific blog post by its ID and sends it as a response. These routes will be used to display the blog posts on the index page and to view individual blog posts.