refactor: Update nav.ejs to fix link for creating blogs
The code changes in this commit update the nav.ejs file to fix the link for creating blogs. The previous link was pointing to "/blogs" which is incorrect. It has been updated to "/create" to correctly direct users to the create blog page. This change aligns with recent user commits that added routes for retrieving all blogs and a single blog, and the addition of the Blog model to the project.