feat: Update create.ejs form action and add name attributes
The code changes in this commit update the `create.ejs` file to fix the form action and add name attributes to the input fields. The form action was previously empty, and it has been updated to `/blogs` to correctly submit the form data to the server. Additionally, the name attributes have been added to the input fields to ensure that the form data can be properly identified and processed on the server side. This change aligns with recent user commits that added routes for creating and retrieving blogs, and the addition of the Blog model to the project.