feat: Add "Read more" link to blog snippets on index page
The code changes in this commit add a "Read more" link to the blog snippets on the index page. This link allows users to navigate to the individual blog post page for more details. This change aligns with recent user commits that added functionality for creating and retrieving blogs, and the addition of the Blog model to the project.